+++ app/src/main/java/com/emplk/realestatemanager/domain/locale_formatting/LocaleFormattingRepository.kt
@@ -8,8 +8,8 @@ interface LocaleFormattingRepository {
     fun convertSquareMetersToSquareFeet(squareMeters: BigDecimal): BigDecimal
     fun getLocale(): Locale
     fun getLocaleSurfaceUnitFormatting(): SurfaceUnitType
-    fun convertDollarToEuro(dollar: BigDecimal, currencyRate: Double): BigDecimal
-    fun convertEuroToDollar(euro: BigDecimal, currencyRate: Double): BigDecimal
+    fun convertDollarToEuro(dollar: BigDecimal, currencyRate: BigDecimal): BigDecimal
+    fun convertEuroToDollar(euro: BigDecimal, currencyRate: BigDecimal): BigDecimal
     fun formatPrice(price: BigDecimal): String
     fun getLocaleCurrencyFormatting(): CurrencyType
 }
